+++ tests/testthat/test-read.R
@@ -10,6 +10,14 @@ test_that("Reading H5AD works", {
     expect_identical(colnames(colData(sce)), "cell_type")
 })
 
+test_that("Reading H5AD works with version 0.7.6", {
+    sce <- readH5AD(file, version = "0.7.6")
+    expect_s4_class(sce, "SingleCellExperiment")
+
+    expect_identical(assayNames(sce), "X")
+    expect_identical(colnames(colData(sce)), "cell_type")
+})
+
 test_that("Reading H5AD works with HDF5Arrays", {
     sce <- readH5AD(file, use_hdf5 = TRUE)
     expect_s4_class(sce, "SingleCellExperiment")